gradle syntax :)
