+++ gradle.properties
@@ -9,6 +9,7 @@ INTELLIJ_IDEA_VERSION=2019.3.4
 JVM_TARGET_VERSION=1.8
 KOTLIN_TEST_VERSION=3.4.2
 ASSERTJ_VERSION=3.13.2
+IDEA_APACHE_ANT_VERSION=1.10.5
 DOKKA_VERSION=0.10.0
 CLASS_GRAPH_VERSION=4.8.47
 JUNIT_VINTAGE_VERSION=5.4.0
