+++ idea-plugin/src/main/kotlin/arrow/meta/ide/testing/dsl/IdeTestSyntax.kt
@@ -5,8 +5,9 @@ import arrow.meta.ide.testing.dsl.icon.IconProviderTestSyntax
 import arrow.meta.ide.testing.dsl.inspection.InspectionTestSyntax
 import arrow.meta.ide.testing.dsl.lineMarker.LineMarkerTestSyntax
 import arrow.meta.ide.testing.env.IdeTestTypeSyntax
-import arrow.meta.ide.testing.env.github.GithubSyntax
+import arrow.meta.ide.testing.env.git.GitSyntax
+import arrow.meta.ide.testing.env.gradle.GradleSyntax
 import arrow.meta.ide.testing.env.resolution.ResolutionSyntax
 
 interface IdeTestSyntax : IdeTestTypeSyntax, LineMarkerTestSyntax, IconProviderTestSyntax,
-  InspectionTestSyntax, ResolutionSyntax, FoldingTestSyntax, GithubSyntax
+  InspectionTestSyntax, ResolutionSyntax, FoldingTestSyntax, GitSyntax, GradleSyntax
