+++ idea-plugin/src/test/java/external/forks/jb/gradle/GradleJpsCompilingTestCase.java
@@ -1,5 +1,5 @@
 // Copyright 2000-2019 JetBrains s.r.o. Use of this source code is governed by the Apache 2.0 license that can be found in the LICENSE file.
-package arrow.meta.ide.jetbrainsGradle;
+package external.forks.jb.gradle;
 
 import com.intellij.openapi.application.ApplicationManager;
 import com.intellij.openapi.compiler.CompileContext;
