+++ idea-plugin/src/test/java/external/forks/jb/gradle/VersionMatcherRule.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package arrow.meta.ide.jetbrainsGradle;
+package external.forks.jb.gradle;
 
 import org.gradle.util.GradleVersion;
 import org.hamcrest.CoreMatchers;
